backend/requirements.txt
flask==2.2.5
mnemonic==0.21
numpy
qrcode
pillow
stegano